+++ extensions/picgo/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Picgo Changelog
 
-## [Bugfix: Plugin Search Error] - 2026-04-01
+## [Image Preview Grid View] - {PR_MERGE_DATE}
+
+- Add a new `Image Grid` view for uploaded image results.
+- Add new preference `Upload Result View`, you can pick the former `Format List` or newly added `Image Grid` view as the default view to display uploaded image results.
+
+## [Bugfix: Plugin Search Error] - 2024-04-01
 
 - Fix error when a plugin package data has no `description` field.